Moved KnownDimensions type to sparse_label_model
