+++ snorkel/labeling/model/label_model.py
@@ -15,7 +15,7 @@ from snorkel.labeling.model.base_labeler import BaseLabeler
 from snorkel.labeling.model.graph_utils import get_clique_tree
 from snorkel.labeling.model.logger import Logger
 from snorkel.types import Config
-from snorkel.types.data import KnownDimensions
+from snorkel.labeling.model.sparse_label_model import KnownDimensions
 from snorkel.utils.config_utils import merge_config
 from snorkel.utils.lr_schedulers import LRSchedulerConfig
 from snorkel.utils.optimizers import OptimizerConfig
